requirements.txt
matplotlib==3.7.4
monai==1.3.0
nibabel==5.3.2
numpy==1.24.1
opencv_python==4.8.1.78
opencv_python_headless==4.9.0.80
pandas==2.0.3
Pillow==9.3.0
Pillow==11.1.0
pycocotools==2.0.7
scikit_learn==1.3.2
skimage==0.0
torch==2.3.0
torchvision==0.18.0
tensorboard==2.11.0